Send flow screen revision
Updates the mega-outdated send flow with the latest from the spending wallet. It includes 13 screens, but does not cover all the various states, as those would be a bit messy and overwhelming. However, all the screens link back to the guide, so people can always reference there.

Also updated the mobile status bar header to use the dynamic island, to go with the times.